+++ noodles-bed/src/io/writer/record/feature_start.rs
@@ -7,8 +7,13 @@ pub(super) fn write_feature_start<W>(writer: &mut W, position: Position) -> io::
 where
     W: Write,
 {
-    let n = usize::from(position) - 1;
-    let mut dst = [0; usize::FORMATTED_SIZE_DECIMAL];
+    // § 1.6.2 "Coordinates: `chromStart`" (2022-01-05): "...`chromStart` must be less than or
+    // equal to 2^64 - 1..."
+    let n = u64::try_from(usize::from(position))
+        .map(|n| n - 1)
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+
+    let mut dst = [0; u64::FORMATTED_SIZE_DECIMAL];
     let buf = lexical_core::write(n, &mut dst);
     writer.write_all(buf)
 }